Houses Page, Logout functionality, AWS DB, Sign out/in button value changes
Added houses to be displayed, created a loading screen while waiting for houses to load. added python script to be ran upon accessing houses page. Added an aws data base to our db file. added logout functionality as well as display username on sign out button and logout for login button if a user is signed in.